pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assistants-framework"
version = "0.6.9"
authors = [
    {name = "Michael Jarvis", email = "nihilok@jarv.dev"},
]
description = "AI Assistant Framework with CLI and Telegram Bot"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["openai", "gpt3", "gpt3.5", "gpt4", "o1", "chatgpt", "chatbot", "assistant", "assistants", "claude", "anthropic", "cli", "telegram", "llm", "bot", "ui", "tui", "coding-assistant", "coding", "programming"]

dependencies = [
    "openai==1.71.0",
    "aiosqlite==0.20.0",
    "loguru==0.7.3",
    "pyperclip==1.9.0",
    "prompt-toolkit==3.0.48",
    "pygments-tsx==1.0.3",
    "pyyaml==6.0.2",
    "anthropic==0.49.0",
    "aiofiles==24.1.0",
    "aiohttp==3.11.11",
    "setproctitle==1.3.5",
    "tiktoken==0.9.0",
    "pydantic~=2.9.2",
]

[project.optional-dependencies]
telegram = [
    "python-telegram-bot==21.10",
]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest==8.3.4",
    "pytest-mock",
]

[project.urls]
"Homepage" = "https://github.com/nihilok/assistants"
"Bug Tracker" = "https://github.com/nihilok/assistants/issues"

[project.scripts]
ai-cli = "assistants.main:main"
ai-tg-bot = "assistants.main_tg:main"
claude = "assistants.claude:main"

[tool.setuptools]
packages = {find = {exclude = ["assistants.tests*"]}}

[tool.pylint]
max-line-length = 100
disable = ["too-many-instance-attributes", "too-many-arguments", "too-many-positional-arguments"]
